+++ src/nvim/autocmd.c
@@ -18,6 +18,7 @@
 #include "nvim/ex_getln.h"
 #include "nvim/fileio.h"
 #include "nvim/getchar.h"
+#include "nvim/insexpand.h"
 #include "nvim/lua/executor.h"
 #include "nvim/map.h"
 #include "nvim/option.h"
