+++ src/nvim/edit.h
@@ -1,26 +1,8 @@
 #ifndef NVIM_EDIT_H
 #define NVIM_EDIT_H
 
+#include "nvim/autocmd.h"
 #include "nvim/vim.h"
-
-/*
- * Array indexes used for cptext argument of ins_compl_add().
- */
-#define CPT_ABBR        0   // "abbr"
-#define CPT_MENU        1   // "menu"
-#define CPT_KIND        2   // "kind"
-#define CPT_INFO        3   // "info"
-#define CPT_COUNT       4   // Number of entries
-
-// values for cp_flags
-typedef enum {
-  CP_ORIGINAL_TEXT = 1,  // the original text when the expansion begun
-  CP_FREE_FNAME = 2,     // cp_fname is allocated
-  CP_CONT_S_IPOS = 4,    // use CONT_S_IPOS for compl_cont_status
-  CP_EQUAL = 8,          // ins_compl_equal() always returns true
-  CP_ICASE = 16,         // ins_compl_equal ignores case
-  CP_FAST = 32,          // use fast_breakcheck instead of os_breakcheck
-} cp_flags_T;
 
 typedef int (*IndentGetter)(void);
