+++ src/nvim/option.c
@@ -49,6 +49,7 @@
 #include "nvim/highlight.h"
 #include "nvim/highlight_group.h"
 #include "nvim/indent_c.h"
+#include "nvim/insexpand.h"
 #include "nvim/keycodes.h"
 #include "nvim/macros.h"
 #include "nvim/mapping.h"
