+++ src/nvim/state.c
@@ -5,10 +5,10 @@
 
 #include "nvim/ascii.h"
 #include "nvim/autocmd.h"
-#include "nvim/edit.h"
 #include "nvim/eval.h"
 #include "nvim/ex_docmd.h"
 #include "nvim/getchar.h"
+#include "nvim/insexpand.h"
 #include "nvim/lib/kvec.h"
 #include "nvim/log.h"
 #include "nvim/main.h"
